Rewrite the backend to be compatible with wasm
* Abstraction of SpScanner into a trait
* Gating of the code with wasm32 target when necessary
