+++ src/backend/mod.rs
@@ -3,10 +3,11 @@ mod backend;
 mod blindbit;
 mod structs;
 
-pub use backend::ChainBackend;
-
 #[cfg(target_arch = "wasm32")]
 pub use backend::ChainBackendWasm;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub use backend::ChainBackend;
 
 pub use structs::*;
