+++ src/scanner/mod.rs
@@ -1,3 +1,344 @@
-mod scanner;
+use std::collections::{HashMap, HashSet};
 
-pub use scanner::SpScanner;
+use anyhow::{Error, Result};
+use bitcoin::{
+    absolute::Height, bip158::BlockFilter, hashes::{sha256, Hash}, 
+    Amount, BlockHash, OutPoint, Txid, XOnlyPublicKey
+};
+use futures::Stream;
+use silentpayments::receiving::Label;
+
+use crate::{
+    backend::{BlockData, FilterData, UtxoData},
+    client::{OwnedOutput, SpClient},
+    updater::Updater,
+};
+
+#[cfg(not(target_arch = "wasm32"))]
+use crate::backend::ChainBackend;
+
+#[cfg(target_arch = "wasm32")]
+use crate::backend::ChainBackendWasm;
+
+/// Trait for scanning silent payment blocks
+/// 
+/// This trait abstracts the core scanning functionality, allowing consumers
+/// to implement it with their own constraints and requirements.
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+pub trait SpScanner {
+    /// Scan a range of blocks for silent payment outputs and inputs
+    /// 
+    /// # Arguments
+    /// * `start` - Starting block height (inclusive)
+    /// * `end` - Ending block height (inclusive)
+    /// * `dust_limit` - Minimum amount to consider (dust outputs are ignored)
+    /// * `with_cutthrough` - Whether to use cutthrough optimization
+    async fn scan_blocks(
+        &mut self,
+        start: Height,
+        end: Height,
+        dust_limit: Amount,
+        with_cutthrough: bool,
+    ) -> Result<()>;
+
+    /// Process a single block's data
+    /// 
+    /// # Arguments
+    /// * `blockdata` - Block data containing tweaks and filters
+    /// 
+    /// # Returns
+    /// * `(found_outputs, found_inputs)` - Tuple of found outputs and spent inputs
+    async fn process_block(
+        &mut self,
+        blockdata: BlockData,
+    ) -> Result<(HashMap<OutPoint, OwnedOutput>, HashSet<OutPoint>)>;
+
+    /// Process block outputs to find owned silent payment outputs
+    /// 
+    /// # Arguments
+    /// * `blkheight` - Block height
+    /// * `tweaks` - List of tweak public keys
+    /// * `new_utxo_filter` - Filter data for new UTXOs
+    /// 
+    /// # Returns
+    /// * Map of outpoints to owned outputs
+    async fn process_block_outputs(
+        &self,
+        blkheight: Height,
+        tweaks: Vec<bitcoin::secp256k1::PublicKey>,
+        new_utxo_filter: FilterData,
+    ) -> Result<HashMap<OutPoint, OwnedOutput>>;
+
+    /// Process block inputs to find spent outputs
+    /// 
+    /// # Arguments
+    /// * `blkheight` - Block height
+    /// * `spent_filter` - Filter data for spent outputs
+    /// 
+    /// # Returns
+    /// * Set of spent outpoints
+    async fn process_block_inputs(
+        &self,
+        blkheight: Height,
+        spent_filter: FilterData,
+    ) -> Result<HashSet<OutPoint>>;
+
+    /// Get the block data stream for a range of blocks
+    /// 
+    /// # Arguments
+    /// * `range` - Range of block heights
+    /// * `dust_limit` - Minimum amount to consider
+    /// * `with_cutthrough` - Whether to use cutthrough optimization
+    /// 
+    /// # Returns
+    /// * Stream of block data results
+    fn get_block_data_stream(
+        &self,
+        range: std::ops::RangeInclusive<u32>,
+        dust_limit: Amount,
+        with_cutthrough: bool,
+    ) -> std::pin::Pin<Box<dyn Stream<Item = Result<BlockData>> + Send>>;
+
+    /// Check if scanning should be interrupted
+    /// 
+    /// # Returns
+    /// * `true` if scanning should stop, `false` otherwise
+    fn should_interrupt(&self) -> bool;
+
+    /// Save current state to persistent storage
+    fn save_state(&mut self) -> Result<()>;
+
+    /// Record found outputs for a block
+    /// 
+    /// # Arguments
+    /// * `height` - Block height
+    /// * `block_hash` - Block hash
+    /// * `outputs` - Found outputs
+    fn record_outputs(
+        &mut self,
+        height: Height,
+        block_hash: BlockHash,
+        outputs: HashMap<OutPoint, OwnedOutput>,
+    ) -> Result<()>;
+
+    /// Record spent inputs for a block
+    /// 
+    /// # Arguments
+    /// * `height` - Block height
+    /// * `block_hash` - Block hash
+    /// * `inputs` - Spent inputs
+    fn record_inputs(
+        &mut self,
+        height: Height,
+        block_hash: BlockHash,
+        inputs: HashSet<OutPoint>,
+    ) -> Result<()>;
+
+    /// Record scan progress
+    /// 
+    /// # Arguments
+    /// * `start` - Start height
+    /// * `current` - Current height
+    /// * `end` - End height
+    fn record_progress(&mut self, start: Height, current: Height, end: Height) -> Result<()>;
+
+    /// Get the silent payment client
+    fn client(&self) -> &SpClient;
+
+    /// Get the chain backend
+    #[cfg(not(target_arch = "wasm32"))]
+    fn backend(&self) -> &dyn ChainBackend;
+
+    /// Get the chain backend (WASM version)
+    #[cfg(target_arch = "wasm32")]
+    fn backend(&self) -> &dyn ChainBackendWasm;
+
+    /// Get the updater
+    fn updater(&mut self) -> &mut dyn Updater;
+
+    // Helper methods with default implementations
+
+    /// Process multiple blocks from a stream
+    /// 
+    /// This is a default implementation that can be overridden if needed
+    async fn process_blocks(
+        &mut self,
+        start: Height,
+        end: Height,
+        block_data_stream: impl Stream<Item = Result<BlockData>> + Unpin + Send,
+    ) -> Result<()> {
+        use futures::StreamExt;
+        use std::time::{Duration, Instant};
+
+        let mut update_time = Instant::now();
+        let mut stream = block_data_stream;
+
+        while let Some(blockdata) = stream.next().await {
+            let blockdata = blockdata?;
+            let blkheight = blockdata.blkheight;
+            let blkhash = blockdata.blkhash;
+
+            // stop scanning and return if interrupted
+            if self.should_interrupt() {
+                self.save_state()?;
+                return Ok(());
+            }
+
+            let mut save_to_storage = false;
+
+            // always save on last block or after 30 seconds since last save
+            if blkheight == end || update_time.elapsed() > Duration::from_secs(30) {
+                save_to_storage = true;
+            }
+
+            let (found_outputs, found_inputs) = self.process_block(blockdata).await?;
+
+            if !found_outputs.is_empty() {
+                save_to_storage = true;
+                self.record_outputs(blkheight, blkhash, found_outputs)?;
+            }
+
+            if !found_inputs.is_empty() {
+                save_to_storage = true;
+                self.record_inputs(blkheight, blkhash, found_inputs)?;
+            }
+
+            // tell the updater we scanned this block
+            self.record_progress(start, blkheight, end)?;
+
+            if save_to_storage {
+                self.save_state()?;
+                update_time = Instant::now();
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Scan UTXOs for a given block and secrets map
+    /// 
+    /// This is a default implementation that can be overridden if needed
+    async fn scan_utxos(
+        &self,
+        blkheight: Height,
+        secrets_map: HashMap<[u8; 34], bitcoin::secp256k1::PublicKey>,
+    ) -> Result<Vec<(Option<Label>, UtxoData, bitcoin::secp256k1::Scalar)>> {
+        let utxos = self.backend().utxos(blkheight).await?;
+
+        let mut res: Vec<(Option<Label>, UtxoData, bitcoin::secp256k1::Scalar)> = vec![];
+
+        // group utxos by the txid
+        let mut txmap: HashMap<Txid, Vec<UtxoData>> = HashMap::new();
+        for utxo in utxos {
+            txmap.entry(utxo.txid).or_default().push(utxo);
+        }
+
+        for utxos in txmap.into_values() {
+            // check if we know the secret to any of the spks
+            let mut secret = None;
+            for utxo in utxos.iter() {
+                let spk = utxo.scriptpubkey.as_bytes();
+                if let Some(s) = secrets_map.get(spk) {
+                    secret = Some(s);
+                    break;
+                }
+            }
+
+            // skip this tx if no secret is found
+            let secret = match secret {
+                Some(secret) => secret,
+                None => continue,
+            };
+
+            let output_keys: Result<Vec<XOnlyPublicKey>> = utxos
+                .iter()
+                .filter_map(|x| {
+                    if x.scriptpubkey.is_p2tr() {
+                        Some(
+                            XOnlyPublicKey::from_slice(&x.scriptpubkey.as_bytes()[2..])
+                                .map_err(Error::new),
+                        )
+                    } else {
+                        None
+                    }
+                })
+                .collect();
+
+            let ours = self
+                .client()
+                .sp_receiver
+                .scan_transaction(secret, output_keys?)?;
+
+            for utxo in utxos {
+                if !utxo.scriptpubkey.is_p2tr() || utxo.spent {
+                    continue;
+                }
+
+                match XOnlyPublicKey::from_slice(&utxo.scriptpubkey.as_bytes()[2..]) {
+                    Ok(xonly) => {
+                        for (label, map) in ours.iter() {
+                            if let Some(scalar) = map.get(&xonly) {
+                                res.push((label.clone(), utxo, *scalar));
+                                break;
+                            }
+                        }
+                    }
+                    Err(_) => todo!(),
+                }
+            }
+        }
+
+        Ok(res)
+    }
+
+    /// Check if block contains relevant output transactions
+    /// 
+    /// This is a default implementation that can be overridden if needed
+    fn check_block_outputs(
+        created_utxo_filter: BlockFilter,
+        blkhash: BlockHash,
+        candidate_spks: Vec<&[u8; 34]>,
+    ) -> Result<bool> {
+        // check output scripts
+        let output_keys: Vec<_> = candidate_spks
+            .into_iter()
+            .map(|spk| spk[2..].as_ref())
+            .collect();
+
+        // note: match will always return true for an empty query!
+        if !output_keys.is_empty() {
+            Ok(created_utxo_filter.match_any(&blkhash, &mut output_keys.into_iter())?)
+        } else {
+            Ok(false)
+        }
+    }
+
+    /// Get input hashes for owned outpoints
+    /// 
+    /// This is a default implementation that can be overridden if needed
+    fn get_input_hashes(&self, blkhash: BlockHash) -> Result<HashMap<[u8; 8], OutPoint>> {
+        let mut map: HashMap<[u8; 8], OutPoint> = HashMap::new();
+
+        // This method needs access to owned_outpoints, which should be provided by the implementor
+        // For now, we'll return an empty map - implementors should override this method
+        Ok(map)
+    }
+
+    /// Check if block contains relevant input transactions
+    /// 
+    /// This is a default implementation that can be overridden if needed
+    fn check_block_inputs(
+        &self,
+        spent_filter: BlockFilter,
+        blkhash: BlockHash,
+        input_hashes: Vec<[u8; 8]>,
+    ) -> Result<bool> {
+        // note: match will always return true for an empty query!
+        if !input_hashes.is_empty() {
+            Ok(spent_filter.match_any(&blkhash, &mut input_hashes.into_iter())?)
+        } else {
+            Ok(false)
+        }
+    }
+}
